requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
Jinja2==3.1.2
requests==2.31.0
asyncpg==0.29.0
aiofile==3.8.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0